// src/generated/model/userAdminModel.ts
/**
 * Datafood Server API
 * API contract of Datafood server application BRASIDAS
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RoleModel } from './roleModel';


export interface UserAdminModel { 
    /**
     * id of the user
     */
    id?: string;
    /**
     * email of the user
     */
    email?: string;
    /**
     * name of the user
     */
    name?: string;
    /**
     * firstname of the user
     */
    firstname?: string;
    /**
     * username of the user
     */
    username?: string;
    /**
     * password of the user
     */
    password?: string;
    /**
     * image url of the user
     */
    image?: string;
    /**
     * google id of the user
     */
    idGoogle?: string;
    /**
     * boolean whether user is active or not
     */
    isActive?: boolean;
    /**
     * username of the user
     */
    sessionExpiration?: number;
    /**
     * height of the user
     */
    height?: number;
    /**
     * weight of the user
     */
    weight?: number;
    /**
     * gender id of the user
     */
    gender?: string;
    /**
     * Date de création du document
     */
    dateCreation?: string;
    /**
     * origin id of the user
     */
    origin?: string;
    /**
     * user roles
     */
    roles?: Array<RoleModel>;
    /**
     * user roles as strings
     */
    rolesStringList?: Array<string>;
}